requirements.txt
pillow
opencv-python
ultralytics
rfdetr
